#404 : 김혜원 : mystudy : Test.js 파일 변경
